# docs/api/openapi-completo.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Proyecto Arquitectura - API Unificada",
    "description": "API Gateway que expone los microservicios del ecosistema de e-commerce. Incluye servicios de gestion de ordenes (Order Service), catalogo de productos (Catalog Service) con cache Redis, y busqueda full-text (Search) con Elasticsearch.\n\n## Arquitectura\n- **API Gateway**: Punto de entrada centralizado (Spring Cloud Gateway)\n- **Order Service**: Gestion de ordenes con notificaciones via SQS/Lambda\n- **Catalog Service**: CRUD de productos con cache Redis\n- **Search Service**: Busqueda full-text con Elasticsearch\n\n## Tecnologias\n- Spring Boot 3.x\n- PostgreSQL (persistencia)\n- Redis (cache)\n- Elasticsearch (busqueda)\n- AWS SQS + Lambda (mensajeria serverless)\n- Docker (contenedores)",
    "version": "1.0.0",
    "contact": {
      "name": "Equipo de Desarrollo",
      "email": "dev@proyecto.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "API Gateway (produccion local)"
    },
    {
      "url": "http://localhost:8081",
      "description": "Order Service (desarrollo directo)"
    },
    {
      "url": "http://localhost:8082",
      "description": "Catalog Service (desarrollo directo)"
    }
  ],
...
  "paths": {
    "/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Obtener todas las ordenes",
        "description": "Retorna una lista con todas las ordenes registradas en el sistema",
        "operationId": "getAllOrders",
        "responses": {
          "200": {
            "description": "Lista de ordenes obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                },
                "example": [
                  {
                    "id": 1,
                    "customerName": "Juan Perez",
                    "createdAt": "2024-01-15T10:30:00",
                    "status": "CREATED",
                    "total": 150.50
                  },
                  {
                    "id": 2,
                    "customerName": "Maria Garcia",
                    "createdAt": "2024-01-15T11:45:00",
                    "status": "NOTIFIED",
                    "total": 299.99
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Orders"],
        "summary": "Crear nueva orden",
        "description": "Crea una nueva orden y publica un evento a la cola SQS para notificacion asincrona via Lambda",
        "operationId": "createOrder",
        "requestBody": {
          "description": "Datos de la orden a crear",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              },
              "example": {
                "customerName": "Juan Perez",
                "total": 150.50
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Orden creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                },
                "example": {
                  "id": 1,
                  "customerName": "Juan Perez",
                  "createdAt": "2024-01-15T10:30:00",
                  "status": "CREATED",
                  "total": 150.50
                }
              }
            }
          },
          "400": {
            "description": "Datos de orden invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Obtener orden por ID",
        "description": "Busca y retorna una orden especifica por su identificador",
        "operationId": "getOrderById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Orden no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Not Found",
                  "message": "Orden con ID 999 no encontrada"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "tags": ["Orders"],
        "summary": "Actualizar estado de orden",
        "description": "Actualiza el estado de una orden existente. Usado por Lambda para marcar como NOTIFIED despues de procesar el mensaje SQS.",
        "operationId": "updateOrderStatus",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la orden",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "description": "Nuevo estado de la orden",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusUpdate"
              },
              "example": {
                "status": "NOTIFIED"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Estado invalido o no proporcionado"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "description": "Entidad que representa una orden de compra",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "ID unico de la orden",
            "readOnly": true,
            "example": 1
          },
          "customerName": {
            "type": "string",
            "description": "Nombre del cliente",
            "example": "Juan Perez"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora de creacion de la orden",
            "readOnly": true,
            "example": "2024-01-15T10:30:00"
          },
          "status": {
            "type": "string",
            "description": "Estado actual de la orden",
            "enum": ["CREATED", "NOTIFIED", "PROCESSING", "COMPLETED"],
            "example": "CREATED"
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Total de la orden en dolares",
            "minimum": 0,
            "example": 150.50
          }
        },
        "required": ["customerName", "total"]
      },
      "OrderCreate": {
        "type": "object",
        "description": "Datos para crear una nueva orden",
        "properties": {
          "customerName": {
            "type": "string",
            "description": "Nombre del cliente",
            "example": "Juan Perez"
          },
          "total": {
            "type": "number",
            "format": "double",
            "description": "Total de la orden en dolares",
            "minimum": 0,
            "example": 150.50
          }
        },
        "required": ["customerName", "total"]
      },
      "StatusUpdate": {
        "type": "object",
        "description": "Datos para actualizar el estado de una orden",
        "properties": {
          "status": {
            "type": "string",
            "description": "Nuevo estado de la orden",
            "enum": ["CREATED", "NOTIFIED", "PROCESSING", "COMPLETED"],
            "example": "NOTIFIED"
          }
        },
        "required": ["status"]
      },
...
      "Error": {
        "type": "object",
        "description": "Respuesta de error",
        "properties": {
          "error": {
            "type": "string",
            "description": "Tipo de error",
            "example": "Not Found"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error",
            "example": "Recurso no encontrado"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha y hora del error"
          }
        }
      }
    }
  }
}